requirements.txt
bigdl-llm[xpu]==2.4.0
peft==0.5.0
accelerate==0.23.0
transformers==4.34.0 
datasets==2.15 
fsspec==2023.9.2
accelerate==0.23.0
datasets==2.15.0
huggingface-hub==0.17.3
matplotlib==3.7.1
neural-compressor==2.2
numpy==1.24.3
opencv-python==4.9.0.80
pandas==2.0.3
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
tokenizers==0.14.1
torch==2.0.1a0+cxx11.abi
torchvision==0.15.2a0+cxx11.abi
tornado==6.3.2
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.9.0
wandb==0.15.12
